Fixes tests and mid-calculation Clear button
The Clear button was not just resetting the input mid calculation.
When adding a test to fix this, I noticed the other tests had
broken due to clamping and formatting logic.  I updated the tests
and added a default clamp check, as well as the text to confirm
mid calculation clear behavior and fixing the underlying issue.