Updates the view resizing calculation to be based on height.
Before the extent of resizing scenarios was understood, we
selected the narrowest dimension for sizing the interface fonts
as it was easier to ensure that the fonts wouldn't exceed the
over interface size, but for the tall layout, this lead to
the fonts appearing smaller than ideal.

After splitting the layout handling into 3 different modes, it
became easier to tweak the per-layout constants to get good
sizing, so we moved the layout to always use height so that
the tall layouts had better button sizing.