Adds key clicks for the buttons.
We are not a custom input or keyboard, so we cannot use the
playInputClick API, though we are loading the standard system
sounds.  As such, we do not respect the keyboard click system
setting.  However, we do respect silent mode.

A setting may be added in future to control this.