Preloads sound data for key clicks.
The sounds are further shortened and quieted to try to avoid
clipping the sounds, which can lead to an ugly crackle.  The
sound volume playback is also set lower, and if there is a
sound playing, it is quickly faded out rather than just stopping
it.  Finally, we preload the data for the sounds from storage
to memory to reduce the sound start time, and also to allow
us to create a new player quickly for each sound, to allow any
previous sound to wrap up as we start the new.